+++ cmd/internal/cli/verify.go
@@ -29,6 +29,7 @@ var (
 	certificatePath              string // --certificate flag
 	certificateIntermediatesPath string // --certificate-intermediates flag
 	certificateRootsPath         string // --certificate-roots flag
+	ocspVerify                   bool   // --ocsp-verify flag
 	pubKeyPath                   string // --key flag
 	localVerify                  bool   // -l flag
 	jsonVerify                   bool   // -j flag
@@ -117,6 +118,16 @@ var verifyCertificateRootsFlag = cmdline.Flag{
 	EnvKeys:      []string{"VERIFY_ROOTS"},
 }
 
+// --ocsp-verify
+var verifyOCSPFlag = cmdline.Flag{
+	ID:           "ocspVerifyFlag",
+	Value:        &ocspVerify,
+	DefaultValue: false,
+	Name:         "ocsp-verify",
+	Usage:        "enable online revocation check for certificates",
+	EnvKeys:      []string{"VERIFY_OCSP"},
+}
+
 // --key
 var verifyPublicKeyFlag = cmdline.Flag{
 	ID:           "publicKeyFlag",
@@ -179,6 +190,7 @@ func init() {
 		cmdManager.RegisterFlagForCmd(&verifyCertificateFlag, VerifyCmd)
 		cmdManager.RegisterFlagForCmd(&verifyCertificateIntermediatesFlag, VerifyCmd)
 		cmdManager.RegisterFlagForCmd(&verifyCertificateRootsFlag, VerifyCmd)
+		cmdManager.RegisterFlagForCmd(&verifyOCSPFlag, VerifyCmd)
 		cmdManager.RegisterFlagForCmd(&verifyPublicKeyFlag, VerifyCmd)
 		cmdManager.RegisterFlagForCmd(&verifyLocalFlag, VerifyCmd)
 		cmdManager.RegisterFlagForCmd(&verifyJSONFlag, VerifyCmd)
@@ -230,6 +242,10 @@ func doVerifyCmd(cmd *cobra.Command, cpath string) {
 				sylog.Fatalf("Failed to load root certificates: %v", err)
 			}
 			opts = append(opts, apptainer.OptVerifyWithRoots(p))
+		}
+
+		if cmd.Flag(verifyOCSPFlag.Name).Changed {
+			opts = append(opts, apptainer.OptVerifyWithOCSP())
 		}
 
 	case cmd.Flag(verifyPublicKeyFlag.Name).Changed: